Mainly worked on infoscreen.js scrolling and API call timeouts
Now there is a list of global variables at the beginning of infoscreen.js that specify how often
weather data, food data and event data should be updated. Also implemented an expiration mechanism
for slides. Models.py includes an expires_at field that is used as a classname in the slide
templates. This class is then parsed in the function scroll() and if it exceed the current timestamp
the slide is removed.